Moved module css to imports on base.scss, see longer desc.
I changed the module css to be partials and include them in the base.scss file. This is to prevent from having to include bootstrap in each file and having it compile with it resulting in hella bloat.

Another way is to look into CSSO for Gulp to remove the duplication from the final file; which i'll explore.